chore: move slow tests to integration tests
